+++ src/graphql/index.ts
@@ -1,3 +1,5 @@
 export * from "./User"
 export * from "./Auth"
 export * from "./Category"
+export * from "../queries"
+export * from "../mutation"
